plugins/sysprof: port to GTK 4
This port is incomplete. It does not yet handle IdeRunTool to the point of
using a subprocess to spawn the worker. That will be needed for this to
be complete.

We might be able to expand sysprof-cli to do that in a more programmable
fashion without having to create a new executable.