plugins/testui: port to GTK 4
This also breaks things out into a new panel. The position may move still
but this at least gets it into shape to be used.

There are still issues to be ironed out, like why our PTY breaks after
running the first test. We also want to add a context menu so that we
can allow running under the debugger.