plugins/flatpak: apply prepend/append path from manifests
Honestly, this really ought to be in IdePipeline, but that will require
a new vfunc from the runtimes to get their "default path" so that the
flatpak one can return "/app/bin:/usr/bin".

Hopefully can do that as a followup. But this gets it so I can configure
Builder from Builder again.