libide-gui: minimal search popover display
Still lots to do here, but we just need something minimal now to be able
to ship something semi-functional. This displays the results in a list
view.

Some day I'd like to have sections and grouping, but that will have to
wait until we have new features from listview for it.