libide-foundry: log file removal during rebuild
Before we can rebuild, we potentially have to remove a lot of files. This
is unfortunately slower than it could be from our reaper implementation,
but we should at least log what we're doing even if it's a bit slow so that
users at least know something is going on.