libide/foundry: add ide_pty_new_sync()
This is to unify the creation of PTY for use within Builder so that we
don't have to track what has which sort of options and where. The goal is
to have a PTY that is suitable for passing across to another PTY namespace
(which requires further work in IdeRunContext).