libide/foundry: use PTY intercept when running tests
We will need this so that we can eventually parse unit test output for the
various kinds of providers (tap, etc). But furthermore, it actually allows
us to have multiple processes share a VtePty which is then rendered within
the testui VteTerminal widget.

Also make the IdeTest run_async() API private so that we can adjust it as
necessary. The public API is to run via the test-manager.