libide/search: add unload vfunc and remove context parameter
The are IdeObjects now so they can just get the context from the IdeObject
rather than the parameter.

Unfortunately, we can't guarantee IdeContext object at unload time, so we
won't pass a parameter to that anyway. Things that need it should keep a
reference to what they use as members instead.