libide/search: wrap search engine results in IdeSearchResults
This will allow us to do followup filtering without having to requery
the individual search providers. That's useful as currently we have to
inflate all the search results anyway, and so we'd like to avoid doing
that on most keypresses.

Instead, the popover (in a followup commit) will be able to simply
refilter the previous set and only if it fails, discard the result.