libide/threading: helper to send signal upon cancellation
This allows us to easily chain a GCancellable::cancelled signal emission
into a signal() to a subprocess. Use as such:

  ide_subprocess_send_signal_upon_cancel(subprocess, cancellable, SIGKILL);

This uses weak refs to avoid any additional cycles on the objects and
attempts to stop the connection based on where the reference is dropped
(via GObject data list or GCancellable signal connection).