libide/search: raise per-provider max search results
Now that we have GtkListView, we can raise this a bit to both stay
interactive but also try to get a reasonable set that is still browsable
by the user.

We might, at some point, add a toggle for an unlimited result set, but
I don't think that is necessary now.