plugins/buildui: move run command selection to preferences
This really belongs in project settings rather than a custom dialog that
is completely separate from project settings.

It would be nice to use AdwComboRow, but we need search for this to be
useful for projects like GTK. Especially since the DropDown looks bad
even with "flat" on the row.

We still probably want to add a switch row so that we can allow the default
selection via providers to work.