plugins/sysprof: setup control channel for gnome-builder-sysprof
The goal here is for Builder to control this intermediate process and be
able to deliver signals/force-exit through the intermediate to the target
subprocess. Additionally, it can get some information about what is
going on which we could proxy to IdeNotification at some point.

We still need the Builder-side of this in GbpSysprofTool.