plugins/sysprof: use sysprof-agent to profile processes
gnome-builder-sysprof is useful enough generally to get into SDKs that I've
moved it into Sysprof directly. That way we can hopefully do more things
like memprof on GNOME apps if we get it in gnome-build-meta.

This switches to sysprof-agent and sets up the proper D-Bus API for it
to work instead of our little internal thing.

If we don't find sysprof-agent in the runtime, we just use our bundled
version within the Builder file-system namespace (which limits us, but
works the same as this did previously).