libide/foundry: remove assertion
We need this to be possible, since we could have a situation where we
need to prepare a runtime on the host which may not match what is in
the pipeline (like locating/running LSPs/tooling from the host).

Just debug log the instance instead and use it for troubleshooting.