libide/core: add ide_context_register_settings()
The goal here is to make it easy to register settings that can be changed
for both the application and the project. Readback of project settings
always falls through to the lower layers but writes to the project
overrides. For the application it will write and write to the app-wide
gsettings.