libide/core: simplify IdeSettings
We always use org.gnome.builer. as a prefix (except for org.gnome.builder)
so we can codify that into our requirements here. That makes things a lot
easier to use since we can drop all this relative path madness and
discover app vs project settings via the presence of the schema path.