libide/gui: move workbench to IdeActionMixin
Also, add our workbench action group to the context so that we can resolve
actions for the workbench through it.

The IdeActionMuxer takes a weak pointer on the instance which is nice in
that it allows us indirection from the action group even when installing
into the widget hierarchy. That's actually better than what we had
previously with workbnech being *the* action-group as it can make
lifecycle a bit annoying.