plugins/c-pack: Re-enable C indenter
Rewrite a good chunk of the C indenter to work with GSV5 APIs, and
re-enable it. Most of the diff is composed or shuffling functions
around, and replacing `GtkTextIter *begin, *end` to simply `*iter`.
Other than that, the core logic is about the same, so it shouldn't
require re-learning how the indenter works entirely.