plugins/editorui: use IdeActionMixin/IdeActionMuxer
This cleans up things a bit so we don't have to use a bunch of
GPropertyAction for each property. Instead use a mixin/muxer for the
addin and then sub-IdePropertyActionGroup for the collection of properties
and actions we care about.

Additionally, this reduces the amount of weak refs we have to just the
two for IdePropertyActionGroup:item instead of one-per-action.