libide/tweaks: start fleshing out tweaks engine
Okay, so the goal here is to make a more descriptive way to do our
preferences than the madness we have right now. It's just not sustainable
and I'm sitting here loating the process of going through the rest of them
and shoehorning them in. So we need something better.

We need a way to describe various things, attach them to settings, and
somewhat dynamically do so since sub-pages can alter the path to settings
we might be interested in.

Additionally, we need the ability to merge plugins definitions together
dynamically like we do for menus.

Also, like popover menus, we want the ability to create widgets on demand
for where declarative things break down. The light/dark switcher, for
example, would never work in a declarative manner.

We also want to be able to search for settings nicely and elevate them
to the global searech provider too for quick tweaking.

Ultimtaely, these will all map to libadwaita preference pages, groups,
rows, etc. But the "schematic" if you will, will live in auto-merged
.ui files in plugins.

I'm not going the GMenu route because my quick prototype there seemed to
indicate I would spend a lot of time working around menu structures. And
quite frankly, GMenuModel is madness, much like GActionGroup and I already
have enough pain in that area.