libide/sourceview: allow 3/4 overscroll of source view
We used to have this as an option in the old sourceview, but it is not
worth the effort to be that precise. Just implement something reasonable
up-front and then we can delete the option going forward. It's just too
much of an annoyance to track that level of detail/font-measurements/etc
just so you can choose to have X pixels of overscroll.