libide/tweaks: make tests more automated
This moves to something a bit more like refspec, that allows us to have
a number of tests for UI merging and then compare to the expected output.

Of course, we'll have to update these a bunch while we implement the core
types, but it's a good start for what we need.