plugins/newcomers: embed symbolic pngs instead of installing them
We want this plugin to load rather quickly, so ensuring that the icons we
will need are not only pre-compiled (symbolic png) but also loaded in the
same binary executable ensures sequential data loading from disk.

We avoid complex loaders (like librsvg) and save IOPS.

The odd thing about this though is that we can't use the same naming or
paths that we would with hicolor on disk.

 - Need to put it in actions/ for some reason
 - Can't use -symbolc.symbolic.png, just -symbolic.png
 - Make sure we have the 32@1x and 32@2x (64x64) variants
 - Don't install the SVG to causing it to potentially be preferred