libide/tweaks: rename to IdeTweaksFactory
We can make this a bit more generic and use it in a few more places so
that factory code need not be duplicated.

Additionally, start on some private API to inflate items from the factory.
We still need deep copy support in IdeTweaksItem so that we can snapshot
the child of the factory.

This will get used from IdeTweaksModel which will give us a GListModel of
all the inflated items from an IdeTweaksItem including factory generated
children (which do not get attached to the graph).