libide/tweaks: allow settings to be defined in groups
They still get attached to the page so we have a single muxer, but this
at least makes it nicer to define in .ui files without extending the
internal-child directly.

There is possibility for clash if you had app vs project in the same
page, but that should definitely be avoided from a UI standpoint anyway.