libide/core: add property macros
This adds various macros we can use for string/boolean properties which
are really common and annoying to type every time, let alone maintain.

There will surely be a better way to do this in the future, but this at
least cuts down the boilerplate a bit.

If you want a project for GNOME 44, come modernize the libide/plugin code
to use these instead.