libide/tweaks: add binding abstraction
Rather than all the settings/object binding madness we have now, I want
to see IdeTweaksWidget get a "binding" property that can have these as
an object defined within it.

That will allow for us to do a lot more interesting things without having
to shoehorn the widgets multiple property bindings and/or actions.

I don't like how many GSettings are created here, but we can optimize that
later if it becomes a problem. GSettings already has backends which make
a lot of that "fast enough".