libide/tweaks: use IdeTweaksBinding from IdeTweaksDirectory
And port the various consumers over. It's a bit more lines of XML from the
definitions, but it's much cleaner to maintain (and probably easier to
test) from the abstraction standpoint.

We can always add a custom XML element for <setting/> instead if it
becomes problematic to port over.

The other part to look at here is if we want to start caching the GSetting
instances on the root tweaks object. Probably not worth it for now.