prefs: remove legacy preferences/configview
This is all replaced now by libide-tweaks which vastly simplifies how we
can create preferences for settings and objects.

Not quite everything has been brought over, and not everything will be
brought over, but the majority is there. The main missing parts are still
related to configuration editing and toolchain/sysroot configuration and
that is a bit up in the air.