plugins/vcsui: disable branch/tags list
We only want the label for now, since we aren't actually providing the
user with anything if they click on the item.

Of course there are aspirations to do that eventually, it's just not there
yet and we are running out of precious time for 43.

Somebody come save my sanity and implement git workflows for Builder.