build-aux: add variant-schema-compiler
This is handy for us to start using when we pass large variants around
and crack them open as arrays. We can probably start using it for our
search indexes eventually, but for now, I'm going to use it to get our
overhead down when filtering Clang completion results client-side.