plugins/clang: avoid g_variant_lookup()/g_variant_get_child()
This uses variant-schema-compiler to avoid all of the reference heavy
operations with GVariant. Since our usage of GVariant is generally around
one large allocation/GVariant, and then cracking into regions within that
GVariant, we can often forgo all of the intermediate structures and ref
counting associated. We just need to reference the primary variant.

This drops the overhead of gtk_source_completion_queued_update_cb() due
to ide_clang_completion_provider_refilter() from 2.81% to .23% which is
safely in the "this is fine to do without a worker thread" territory.

The result set that was tested, is for a source file containing GTK and a
number of accessory platform libraries in scope.