plugins/top-match: add plugin to elevate exect matches
When we have an exact match, that is found in a lower completion set, we
really want to see it elevated to the top of the results so that if the
user hits Enter, for example, it does the expected thing (or at least the
least likely to be wrong thing).

We still need a better icon for this than the edit-find-symbolic. It would
also be nice if we could add a separator between the rows but that requires
changes to GtkSourceView.