libide/editor: improve save delegate title/subtitle
We want this to be more like the mockups. We use the project directory
name in the subtitle as the path root just to make things a bit more
obvious. If it's not in the path of the context, then we use collapsed
path names starting from HOME. Finally, give URIs something special so
that it's more obvious where it comes from (ssh://host/ for example).