libide/code: port highlight engine to CjhTextRegion
This tries to avoid using the ever growing textmark region to update and
instead tries to be finer grained with runs of invalid areas. It should
help speed up quite a bit how fast we can react to changes.

However, some additional work to only clear highlights from lines we are
about to process would make this feel a lot better.