libide/code: remove tags as part of higlighter update
Clearing the old tags in one fell swoop is very handy from a performance
perspective, as it means we do less work. However, it is very jarring to
the user because it means that the screen can flash as they are typing.

Instead, clear the old tags as we are scanning words so that we more
minimally damage the screen as highlight indexes are updated and/or
words are typed.