libide/gui: remove maximize call
This wasn't needed as we still restore sizes to/from GSettings for the
cases where we aren't using session data. The original thought was that
we'd always start new projects maximized, and then they rely on session
restore after that. But since using -e for editor mode is common enough,
we can keep the GSetting around for that anyway.