libide/terminal: add ide_terminal_page_new_completed()
This is a convenience function to setup a new terminal page that is
intended for display purposes only such as the contents of a previously
run terminal (ie: previous application session).

We want to keep the state around to be reused, but otherwise interacting
with the terminal is not supported.