plugins/buildui: clear bound model when pipeline changes
We need to be sure that the GtkListBox has access to it's references so
that they can be removed in-case the model changes. Additionally, we don't
want it to continue to react to changes, so just clear the model. Odd that
it wasn't done this way to begin with.

Fixes #956