libide/code: add IdeSymbolNode:display-name
This way we can separate items that have names and their display format
(which is pango markup). The only thing using that was the xml symbol
node, which we can quick fixup.

This fixes showing things like impl<foo> in the symbol tree from languages
that, you know, aren't xml.