plugins/codeui: bridge IdeFormatter on save
This uses the formatter for the buffer to format the document upon save
if the format-on-save GSetting is set. It requires the setting to be
enabled and an IdeFormatter registered for the document syntax.

Additionally, since the buffer does not have access to view settings like
the size to draw indentation, we rely on IdeFileSettings to get those
values. In general, those should be in-sync with the document view since
it uses those to bridge to the display too.

Related #1795