plugins/deviced: bridge run context to to gnome-builder-deviced
There is still a bit more to be desired here, like handling cleanup with
the proper signal propagated to the peer, as well as connecting additional
FDs, or running tools on the remote side.

But that is beyond what deviced is capable of, and should really just be
made as something on top of Bonsai. We could also handle SDK/Runtime
installation when missing with that too.