plugins/clang: use -Wno-unused-function for headers
Rather than use our hack for -Dinline, we can just disable the diagnostic
for headers (but make sure we append it so it takes effect after other
diagnostic values like -Wall).

This makes sure we still warn in C/C++ files, but not headers like the
original issue reported.

Related #1807