libide/foundry: add private API to attach PTY to run context
We can make this public in 44, but for now keep it private to simplify the
backport to 43.

We want to use this to attach PTY to the top-most layer of the run context
so that layers like runtimes can see if there is a PTY required. They might
use that information to set additional feature flags.

For example, GbpPodmanRuntime needs this to determine if --tty should be
used when spawning with `podman exec`.