libide/code: lower priority of release buffer source
We want this to be as low as possible so that other GIdleSource will
complete before this one. Particularly for things like task completion
that will be consuming the buffer. We want the hold release delayed until
those tasks have completed.

But since we can't rely on all async ops calling to complete a task, we
can't do that release in the finish and have to result to some timing
and/or main-context tricks like this.