plugins/git: emit monitor changed when index changes
When we discover the IdeVcs::changed signal is emitted, the buffer change
monitor (for git) also needs to emit IdeBufferChangeMonitor::changed so
that the gutter is re-rendered with the updated state.

This matters more now in GTK 4 since the widgets are more likely to retain
state and not render anything new until forced.

Fixes #1820