plugins/ctags: use IdeObject::destroyed for life-cycle management
This makes it so we don't have to track the service from the utility
classes, nor do we need to manually unregister them or hold extra
references for them which would cause cycles.

Instead, we just watch for both of these object's destroy signal from the
IdeObject superclass and unregister them manually.

Fixes #1811