libide/io: use preferred shell for host $PATH discovery
Previously, we tried to discover the user's preferred shell and host $PATH
environment variables concurrently. That prevents us from being able to use
the users preferred shell to discover $PATH.

The side-effect of this is that we may not pick up environment changes that
are defined in places like ~/.bashrc when bash is the preferred shell.

This fixes that by asynchronously discovering $SHELL, then using that to
run $SHELL -l -c 'echo $PATH' assuming we found a shell that we know
supports -l (for --login shell) and -c (for command). Otherwise, we still
fallback to using /bin/sh for discovery.

This fixes an issue where the podman plugin failed to discover tooling like
rust-analyzer being available within the container because we didn't pick
up environment settings from .bashrc which might give access to rustc,
cargo, rust-analyzer, etc.

Fixes #1822