plugins/meson: extract custom targets installed to bin/
If we introspect and find a target that is "custom" and that gets installed
to $prefix/bin/ (although we don't care about $prefix here, just that it
ends in /bin), then assume it's a program and we can execute it.

This fixes automatic discovery of targets in meson based CLI projects
where external tooling is used. One such example is Rust via Meson+Cargo.