libide/foundry: add _ide_run_context_push_user_shell() private API
This could be public API, but is private currently to simplify backporting
to gnome-builder-43. It is similar to ide_run_context_push_shell() but
requests the user's preferred shell instead.

This should only be used when you know you'd have access to the user's
preferred shell (ie: via the host runtime).