plugins/host: add noop runtime
This allows using the bundled runtime of Builder when running from Flatpak
as a build runtime target for applications. It's simpler than selecting
a Flatpak runtime directly as it does not require trampolining through
the host to the Flatpak runtime.

Mostly just for small testing where you don't have another pet container
setup to be used.

Related #1826