libide/foundry: add translation API to IdeConfig
We really want this at the config level than just on a runtime because
it allows more control over complex resolution. For example, Flatpak
manifests will need to look in SDK extensions, SDKs, Debug runtimes,
the platform runtime, and additionally the staging directory while building
the application.

This is a backport for GNOME 43 and therefore API stays private.