libide/editor: compress minimap a bit
We don't need to follow Text Editor in sizes here, as Text Editor is more
of a generic editor. We expect people writing code here, and a bit more
space has a very practical benefit.

I would expect a slight slowdown in text rendering performance, given that
this will require snapshotting more lines (and snapshotting lines in GSV
requires potentially updating the highlight for that region).

3px suggested by @exalm.