plugins/lua-language-server: add lua-language-server support
You still have to install the language server manually, but this gets
things in place to discover/run the language server.

The most annoying part is that you have to use a wrapper script for it
because the language server requires locating lua files bundled with the
project to be run.

Fixes #49