plugins/spellcheck: fix action group usage
This had mismatched gtk/panel widget actions but additionally we want to
use the rest of the IdePage infrastructure to do the action group
attachment of actions (so they can connect to the frame).

Fix all the actions to use the proper prefix of "spellcheck" to match the
plugin module-name.