libide/lsp: handle workspace/configuration fallback more gracefully
This should be sending an [LSPAny?] back, so do that if we fail to load
a configuration from the subclass/etc.

This really needs to change going forward though, as the clients are
expected to reply to various requests in the params, and we are not
providing those when emitting the signal.