libide/sourceview: implement matching ops using buffer signals
We don't want to be using key controllers to do this work because it can
really get in the way of using things like Vim input methods (or any
input methods for that matter).

This requires a bit more trickery, and I'm sure we can make it better, with
how we're detecting key presses, but it seems to be the one thing that
works in the absence of knowing if we're in GtkTextView from a
"commit-text" phase of a input method via a key-press.

Seems to fix the issue with Vim as well.

Fixes #1851