plugins/rust-analyzer: port to generic LSP plugin
This removes some features, such as the rust-analyzer tweaks addin. But
we want to solve this more generically so it is the same across all of the
language servers (and probably more similar to VS Code).

We'll re-introduce other ways to tweak settings, but it may be more like a
settings.json editing.