build-aux: remove bundled language servers
The goal here is to stop doing this crazy bundling, and instead move the
LSPs into the build containres (via SDK extensions) or user installation
in their build environment.

The other option, should we get there, is for plugins to be bundled as
extensions on Flathub which can install LSPs with their extension.