sourceview: use scrolling for scroll_to_insert()
This still keeps our "jump to insert" behavior for the cases where we
absolutely cannot depend on scrolling because the window has not been
displayed to the point that we can do interpolations in GtkAdjustment
without breaking.

The jump behavior is maintained through a new API function which more
closely matches what should be done.

Fixes #1878