libide/webkit: use PanelWidget actions for web prefix
Rather than having two sets of actions here (which was getting in the way
when reparting pages), use the libpanel way of associating actions for
everything.

This requires binding enable-javascript settings to a new property, but
that is fine as it allows programmatic access anyway to consumers of the
IdeWebkitPage widget.

We can backport this since it doesn't add new API calls, only an additional
property on the page.

Fixes #1883