plugins/flatpak: update remotes on worker thread
This gets it out of the processing loop, which is ideal for keepin callers
fast during initialization. Additionally, this changes things to only
update the peer if we added the remote to our private installation.

What we need to watch out for, is if this will cause issues with resolving
named SDK extensions when loading projects. Because we are at risk of that
failing if we cannot rely on the cached information.

Once we can be reasonably certain this handles things better, we can look
into backporting. But I anticipate some sort of fallout, even if minor
from doing this.

Related #1857