libide/lsp: add dynamic LSP plugin search provider
This adds a search provider for all LSP types that searches against the
contents of "workspace/symbols" call. We currently do not check if that is
supported on the peer, but sort of get away with it nonetheless.

Since this uses lazy binding of clients, the idea is that you have to
actually use something that causes the LSP service to spawn elsewhere for
the search provider to be functional. That way we don't spawn *all* LSPs
just for search capabilities.

This fixes the missing search support for rust-analyzer that was lost when
moving to dynamic LSP plugins.