libide/lsp: try harder to avoid starting lazy bound services
If we have never seen a provider that forces the spawning of a LSP, then
we should not spawn them in response to pipeline invalidation. Otherwise
we risk starting a LSP that should have never been started but did just
becaues the pipeline invalidated.

This likely improves the situation for #1874 among others.