libide: Better scrolling to search results that are out of view
If the next (nb: not the previous) search result is outside our current
view, we want to violate the rule of "do minimal scrolling to get
there": In this case context will be somewhat lost (search result is
outside of the view) and the brain has to reorient. People read from top
to bottom, so they naturally look first at the top of the view for the
selection.

So in this specific case we want to have the selected search result
positioned at the top of the view.

There is a special case where this behavior is a bad idea though: In the
wrap-around case where we scroll down, but actually intended to scroll
up (because we pressed arrow-up key). In this case we actually want the
behavior of "minimal scrolling" again.