plugins/meson-templates: use tabs for C templates
Really, want the tabs post-processed into spaces if that is what the
user has set for the configuration, but we should at least be consistent
across the templates so they're always the same.

This uses tabs, but keeps spaces for alignment so they can look correct
at any width for tabs.