plugins/spellcheck: protect against infinte loop
The old version of this code, while trying to be more robust in using
Pango language breaking, also happened to be quite difficult to protect
against infinite loops.

Instead, this assumes that we're getting the full language character set
from Enchant (which appears to be the case) and breaks on that.

If this turns out to be good enough, we can merge this to 43 as well as
transition back to g-t-e.

Fixes #1893