libide/sourceview: remove search bubbles
I'm tired of maintaining these, and people don't seem to like them so I
guess to the bin they go.

The best way to handle these in the future would be to teach PangoRenderer
to be able to do more than just "rectangles" so we can have a rounded rect
that can be drawn with GskPangoRenderer. It would also need enough
information in GtkTextTags to pass that through.