libide/sourceview: disable matching braces on async paste
This is rather cumbersome in that we don't know when the asynchronous paste
will complete from the function that has been chained up. Instead, we just
keep a note that a paste is happening, and skip the next processing of
insert-text as if we were in a key-press (and we won't be by then because
it will have already gotten the key-release).

Fixes #1870