plugins/debuggerui: add keybindings/menu entries for actions
We can't match some common keybindings from other editors due to clashes
with GTKs internal keybindings, but we can do something with similar
effect from F5-F8 and similar Shift semantics.

This adds new menu entries for discoverability and keybindings to activate
from JSON.

F7 is the visibility of cursor toggle in GtkTextView, but that works
fine with Ctrl+F7 so we'll take that for our own in capture.

Fixes #1942