plugins/flatpak: resolve sdk extensions asynchronously
This will mean that the build pipeline isn't ready initially, but will
become ready once the async operation completes.

We need to provide some status to the user that this is occurring, but
that has not yet been designed. Additionally, we want the run button to
still be able to show it's menu, and it doesn't currently.