lide/core: use IdeObjectClass.destroy for cleanup
This fixes in-tree plugins inheriting from IdeObject to use the destroy
vfunc instead of dispose so that we can enforce object cleanup on the
main thread. We must warn when they are used improperly so they get fixed.

The goal here is to avoid a whole class of issues where objects dispose
and release other objects on a thread where we can introduce data-races
or cascading thread-safety issues.

Currently it just emits a critical since there may be external plugins
doing something we don't want.