libide/foundry: allow build systems to alter source directory
This allows the build system, particularly when a project is opened with
a specific project file (meson.build, Makefile, etc) to alter what should
be considered the source directory for the project.

Build systems will need to support this explicitely by implementing the
vfunc or the previous behavior will be used (the project workdir).