plugins/meson: implement IdeBuildSystem.get_srcdir()
This allows building projects using Meson that exist within a subdirectory
of the top-level git directory.

You must open the project by selectin the toplevel "meson.build" file you
want to build instead of simply the directory as that would get re-rooted
to the directory containing .git.

Fixes #830 in the context of Meson