plugins/project-tree: add "filter" support to tree
This is tree "filter" (it's not actually a filter) which builds upon the
search engine from global search to show files.

Why? Because we already have the contents indexed and that means we can
filter things a lot faster than to actually do a scan of the project
tree on-demand. Particularly because the tree is lazily loaded so that
we can support larger projects.

Additionally, we'd want to eventually be able to use this to search within
the contents of files as we will from global search.