libide-core: only import execinfo.h with IDE_ENABLE_TRACE
this is only used when tracing is enabled below, so only import it when
used.

this fixes the build against musl libc (with tracing disabled), because
musl does not implement backtrace()/execinfo.h. since this functionality
is optional, all we have to do is add the same #ifdef guard to the
include as the actual tracing, and move the definition for it above.