plugins/xml-pack: move parsing off main thread
This uses a libdex fiber to run the operation on the thread pool, then
process the result back on the main thread with a DexBlock to apply the
buffer contents.

Additionally, while we don't have full control over libxml2, we can at
least guard our access to the xmlTreeIndentString global across racing
threads.

Also can simplify the indentation string choice a bit by putting the "\t"
at the head of the list which makes indexing by tab-width passthrough.

To avoid extra string copies of the buffer, we use the
ide_buffer_dup_content() helper which increases the chance we re-use what
was already there during the last edit.