libide/core: use GListModel for user information logs
Previously we just had a GSignal that would emit log information when it
occurred. This doesn't help much during startup because there may not be
anything connected to it (and thus logs are lost).

This instead keeps the messages around in a log and exports that as a
GListModel which we can present to the user.

If we decide to use this much more heavily, we may want to push this out
into a journald style log file w/ can be persistent across runs of the
project, for example.

Fixes #1965 in that we now can provide that information to the user if
they check the messages panel.