plugins/golang: add minimal golang project support
This is pretty basic, despite looking like it does a lot. It tries to open
projects containing a "go.mod" as a golang project. It builds using
"go get -v", "go build -v", and cleans with "go clean -v". To run it
does "go run main.go" which is going to clearly fail if it can't find your
main.go (so keep that in mind).

If GOROOT is not specified, it will use your project dir as the GOROOT.