build: remove sysroot plugin
This never really kept up with all the changes in the foundry and needs
significant work to be useful. Particularly around preferences support
and elevating the feature to users.

For now, we'll just stick to the meson toolchain support and drop this.

If you were somehow finding a way to use this (and I don't think it was
currently possible) then do please holler to let me know and what your
use case is/was.