libide/gui: update accels for dynamic shortcuts
If the user changes the shortcut, we want to update the accel label so
it is replicated in popover menus.

We need to still update the rest of the shortcuts to use this style for
identifying shortcuts and and implement writeback support for user
modified shortcuts. But this gets us much closer to being able to both
modify things for the user and show the user what those modifications
are within menus.