plugins/shortcutui: add new model for overrides
This is the base layer for a new override UI that maps to shortcuts now
that we have alternatives for getting access to what should be in the list.

This isn't used yet, as it needs a bit more, but it compiles and so we can
iterate on it from here.