libide/gui: Navigate search using Up and Down keys
By default, GtkSearchEntry binds Ctrl+G and Ctrl+Shift+G to the
'next-match' and 'previous-match' action signals, respectively.
That's fine, but a bit unergonimic.

There's an argument to be made about whether or not GtkSearchEntry
should bind the Up and Down arrow keys too, but that's up in the
air - it might be reasonable that other apps may not want this
behaviour, and the toolkit shouldn't get in the way.

For now, bind the Up and Down arrow keys to the action signals of
the search entry. This should make navigating through search results
a little more ergonomic.